bb-downloader: Move AsyncTempFile to helpers
- impl AsyncWrite
- Needed to write better tests.
