+++ bb-downloader/src/lib.rs
@@ -8,35 +8,11 @@
 //! - Check if a file is available in cache.
 //! - Uses SHA256 for verifying cached files.
 //! - Optional support to download files without caching.
-//!
-//! # Sample Usage
-//!
-//! ```no_run
-//! #[tokio::main]
-//! async fn main() {
-//!     let downloader = bb_downloader::Downloader::new("/tmp").unwrap();
-//!
-//!     let sha = [0u8; 32];
-//!     let url = "https://example.com/img.jpg";
-//!
-//!     // Download with just URL
-//!     downloader.download(url).await.unwrap();
-//!
-//!     // Check if the file is in cache
-//!     assert!(downloader.check_cache_from_url(url).is_some());
-//!
-//!     // Will fetch directly from cache instead of re-downloading
-//!     downloader.download(url).await.unwrap();
-//!
-//!     // Will fetch directly from cache instead of re-downloading
-//!     assert!(!downloader.check_cache_from_sha(sha).await.is_some());
-//!
-//!     // Will re-download the file
-//!     downloader.download_with_sha(url, sha, None).await.unwrap();
-//!
-//!     assert!(downloader.check_cache_from_sha(sha).await.is_some());
-//! }
-//! ```
+
+
+mod helpers;
+
+use helpers::AsyncTempFile;
 
 #[cfg(feature = "json")]
 use serde::de::DeserializeOwned;
@@ -46,7 +22,7 @@ use std::{
     path::{Path, PathBuf},
     time::Duration,
 };
-use tokio::io::{AsyncSeekExt, AsyncWriteExt};
+use tokio::io::AsyncWriteExt;
 use tokio_stream::StreamExt;
 
 pub use reqwest::IntoUrl;
@@ -185,7 +161,7 @@ impl Downloader {
 
         let mut file = AsyncTempFile::new()?;
         {
-            let mut file = tokio::io::BufWriter::new(&mut file.0);
+            let mut file = tokio::io::BufWriter::new(&mut file);
 
             let response = self
                 .client
@@ -309,31 +285,4 @@ fn sha256_from_path(p: &Path) -> io::Result<[u8; 32]> {
         .expect("SHA-256 is 32 bytes");
 
     Ok(hash)
-}
-
-struct AsyncTempFile(tokio::fs::File);
-
-impl AsyncTempFile {
-    fn new() -> io::Result<Self> {
-        let f = tempfile::tempfile()?;
-        Ok(Self(tokio::fs::File::from_std(f)))
-    }
-
-    async fn persist(&mut self, path: &Path) -> io::Result<()> {
-        let mut f = tokio::fs::File::create(path).await?;
-        self.0.seek(io::SeekFrom::Start(0)).await?;
-
-        tokio::io::copy(&mut self.0, &mut f).await?;
-
-        // Causes errors if not present
-        f.flush().await?;
-
-        Ok(())
-    }
-}
-
-impl From<std::fs::File> for AsyncTempFile {
-    fn from(value: std::fs::File) -> Self {
-        Self(tokio::fs::File::from_std(value))
-    }
 }
